+++ marshmallow_recipe/fields.py
@@ -28,7 +28,7 @@ def str_field(
     strip_whitespaces: bool = False,
     post_load: collections.abc.Callable[[str], str] | None = None,
     required_error: str | None = None,
-    null_error: str | None = None,
+    none_error: str | None = None,
     invalid_error: str | None = None,
     **_: Any,
 ) -> m.fields.Field:
@@ -41,7 +41,7 @@ def str_field(
             **default_fields(m.missing),
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -57,7 +57,7 @@ def str_field(
             post_load=post_load,
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -69,7 +69,7 @@ def str_field(
         **(default_fields(None) if default is dataclasses.MISSING else {}),
         **data_key_fields(name),
         error_messages=build_error_messages(
-            required_error=required_error, null_error=null_error, invalid_error=invalid_error
+            required_error=required_error, none_error=none_error, invalid_error=invalid_error
         ),
     )
 
@@ -82,7 +82,7 @@ def bool_field(
     name: str | None = None,
     validate: ValidationFunc | collections.abc.Sequence[ValidationFunc] | None = None,
     required_error: str | None = None,
-    null_error: str | None = None,
+    none_error: str | None = None,
     invalid_error: str | None = None,
     **_: Any,
 ) -> m.fields.Field:
@@ -93,7 +93,7 @@ def bool_field(
             **default_fields(m.missing),
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -107,7 +107,7 @@ def bool_field(
             validate=validate,
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -117,7 +117,7 @@ def bool_field(
         **(default_fields(None) if default is dataclasses.MISSING else {}),
         **data_key_fields(name),
         error_messages=build_error_messages(
-            required_error=required_error, null_error=null_error, invalid_error=invalid_error
+            required_error=required_error, none_error=none_error, invalid_error=invalid_error
         ),
     )
 
@@ -133,7 +133,7 @@ def decimal_field(
     as_string: bool = True,
     validate: ValidationFunc | collections.abc.Sequence[ValidationFunc] | None = None,
     required_error: str | None = None,
-    null_error: str | None = None,
+    none_error: str | None = None,
     invalid_error: str | None = None,
     **_: Any,
 ) -> m.fields.Field:
@@ -147,7 +147,7 @@ def decimal_field(
             **default_fields(m.missing),
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -163,7 +163,7 @@ def decimal_field(
             validate=validate,
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -176,7 +176,7 @@ def decimal_field(
         **(default_fields(None) if default is dataclasses.MISSING else {}),
         **data_key_fields(name),
         error_messages=build_error_messages(
-            required_error=required_error, null_error=null_error, invalid_error=invalid_error
+            required_error=required_error, none_error=none_error, invalid_error=invalid_error
         ),
     )
 
@@ -189,7 +189,7 @@ def int_field(
     name: str | None = None,
     validate: ValidationFunc | collections.abc.Sequence[ValidationFunc] | None = None,
     required_error: str | None = None,
-    null_error: str | None = None,
+    none_error: str | None = None,
     invalid_error: str | None = None,
     **_: Any,
 ) -> m.fields.Field:
@@ -200,7 +200,7 @@ def int_field(
             **default_fields(m.missing),
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
     elif required:
@@ -212,7 +212,7 @@ def int_field(
             validate=validate,
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
     else:
@@ -222,7 +222,7 @@ def int_field(
             **(default_fields(None) if default is dataclasses.MISSING else {}),
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
     return with_type_checks_on_validated(field, (int, str))
@@ -236,7 +236,7 @@ def float_field(
     name: str | None = None,
     validate: ValidationFunc | collections.abc.Sequence[ValidationFunc] | None = None,
     required_error: str | None = None,
-    null_error: str | None = None,
+    none_error: str | None = None,
     invalid_error: str | None = None,
     **_: Any,
 ) -> m.fields.Field:
@@ -247,7 +247,7 @@ def float_field(
             **default_fields(m.missing),
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -260,7 +260,7 @@ def float_field(
             validate=validate,
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -270,7 +270,7 @@ def float_field(
         **(default_fields(None) if default is dataclasses.MISSING else {}),
         **data_key_fields(name),
         error_messages=build_error_messages(
-            required_error=required_error, null_error=null_error, invalid_error=invalid_error
+            required_error=required_error, none_error=none_error, invalid_error=invalid_error
         ),
     )
 
@@ -283,7 +283,7 @@ def uuid_field(
     name: str | None = None,
     validate: ValidationFunc | collections.abc.Sequence[ValidationFunc] | None = None,
     required_error: str | None = None,
-    null_error: str | None = None,
+    none_error: str | None = None,
     invalid_error: str | None = None,
     **_: Any,
 ) -> m.fields.Field:
@@ -294,7 +294,7 @@ def uuid_field(
             **default_fields(m.missing),
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -307,7 +307,7 @@ def uuid_field(
             validate=validate,
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -317,7 +317,7 @@ def uuid_field(
         **(default_fields(None) if default is dataclasses.MISSING else {}),
         **data_key_fields(name),
         error_messages=build_error_messages(
-            required_error=required_error, null_error=null_error, invalid_error=invalid_error
+            required_error=required_error, none_error=none_error, invalid_error=invalid_error
         ),
     )
 
@@ -331,7 +331,7 @@ def datetime_field(
     validate: ValidationFunc | collections.abc.Sequence[ValidationFunc] | None = None,
     format: str | None = None,
     required_error: str | None = None,
-    null_error: str | None = None,
+    none_error: str | None = None,
     invalid_error: str | None = None,
     **_: Any,
 ) -> m.fields.Field:
@@ -343,7 +343,7 @@ def datetime_field(
             **default_fields(m.missing),
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -357,7 +357,7 @@ def datetime_field(
             format=format,
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -368,7 +368,7 @@ def datetime_field(
         **(default_fields(None) if default is dataclasses.MISSING else {}),
         **data_key_fields(name),
         error_messages=build_error_messages(
-            required_error=required_error, null_error=null_error, invalid_error=invalid_error
+            required_error=required_error, none_error=none_error, invalid_error=invalid_error
         ),
     )
 
@@ -381,7 +381,7 @@ def time_field(
     name: str | None = None,
     validate: ValidationFunc | collections.abc.Sequence[ValidationFunc] | None = None,
     required_error: str | None = None,
-    null_error: str | None = None,
+    none_error: str | None = None,
     invalid_error: str | None = None,
     **_: Any,
 ) -> m.fields.Field:
@@ -392,7 +392,7 @@ def time_field(
             **default_fields(m.missing),
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -405,7 +405,7 @@ def time_field(
             validate=validate,
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -415,7 +415,7 @@ def time_field(
         **(default_fields(None) if default is dataclasses.MISSING else {}),
         **data_key_fields(name),
         error_messages=build_error_messages(
-            required_error=required_error, null_error=null_error, invalid_error=invalid_error
+            required_error=required_error, none_error=none_error, invalid_error=invalid_error
         ),
     )
 
@@ -428,7 +428,7 @@ def date_field(
     name: str | None = None,
     validate: ValidationFunc | collections.abc.Sequence[ValidationFunc] | None = None,
     required_error: str | None = None,
-    null_error: str | None = None,
+    none_error: str | None = None,
     invalid_error: str | None = None,
     **_: Any,
 ) -> m.fields.Field:
@@ -439,7 +439,7 @@ def date_field(
             **default_fields(m.missing),
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -452,7 +452,7 @@ def date_field(
             validate=validate,
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -462,7 +462,7 @@ def date_field(
         **(default_fields(None) if default is dataclasses.MISSING else {}),
         **data_key_fields(name),
         error_messages=build_error_messages(
-            required_error=required_error, null_error=null_error, invalid_error=invalid_error
+            required_error=required_error, none_error=none_error, invalid_error=invalid_error
         ),
     )
 
@@ -476,7 +476,7 @@ def nested_field(
     name: str | None = None,
     validate: ValidationFunc | collections.abc.Sequence[ValidationFunc] | None = None,
     required_error: str | None = None,
-    null_error: str | None = None,
+    none_error: str | None = None,
     invalid_error: str | None = None,
     **_: Any,
 ) -> m.fields.Field:
@@ -488,7 +488,7 @@ def nested_field(
             **default_fields(m.missing),
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -502,7 +502,7 @@ def nested_field(
             validate=validate,
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -513,7 +513,7 @@ def nested_field(
         **(default_fields(None) if default is dataclasses.MISSING else {}),
         **data_key_fields(name),
         error_messages=build_error_messages(
-            required_error=required_error, null_error=null_error, invalid_error=invalid_error
+            required_error=required_error, none_error=none_error, invalid_error=invalid_error
         ),
     )
 
@@ -527,7 +527,7 @@ def list_field(
     name: str | None = None,
     validate: ValidationFunc | collections.abc.Sequence[ValidationFunc] | None = None,
     required_error: str | None = None,
-    null_error: str | None = None,
+    none_error: str | None = None,
     invalid_error: str | None = None,
     **_: Any,
 ) -> m.fields.Field:
@@ -539,7 +539,7 @@ def list_field(
             **default_fields(m.missing),
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -553,7 +553,7 @@ def list_field(
             validate=validate,
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -564,7 +564,7 @@ def list_field(
         **(default_fields(None) if default is dataclasses.MISSING else {}),
         **data_key_fields(name),
         error_messages=build_error_messages(
-            required_error=required_error, null_error=null_error, invalid_error=invalid_error
+            required_error=required_error, none_error=none_error, invalid_error=invalid_error
         ),
     )
 
@@ -578,7 +578,7 @@ def set_field(
     name: str | None = None,
     validate: ValidationFunc | collections.abc.Sequence[ValidationFunc] | None = None,
     required_error: str | None = None,
-    null_error: str | None = None,
+    none_error: str | None = None,
     invalid_error: str | None = None,
     **_: Any,
 ) -> m.fields.Field:
@@ -590,7 +590,7 @@ def set_field(
             **default_fields(m.missing),
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -604,7 +604,7 @@ def set_field(
             validate=validate,
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -615,7 +615,7 @@ def set_field(
         **(default_fields(None) if default is dataclasses.MISSING else {}),
         **data_key_fields(name),
         error_messages=build_error_messages(
-            required_error=required_error, null_error=null_error, invalid_error=invalid_error
+            required_error=required_error, none_error=none_error, invalid_error=invalid_error
         ),
     )
 
@@ -629,7 +629,7 @@ def frozen_set_field(
     name: str | None = None,
     validate: ValidationFunc | collections.abc.Sequence[ValidationFunc] | None = None,
     required_error: str | None = None,
-    null_error: str | None = None,
+    none_error: str | None = None,
     invalid_error: str | None = None,
     **_: Any,
 ) -> m.fields.Field:
@@ -641,7 +641,7 @@ def frozen_set_field(
             **default_fields(m.missing),
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -655,7 +655,7 @@ def frozen_set_field(
             validate=validate,
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -666,7 +666,7 @@ def frozen_set_field(
         **(default_fields(None) if default is dataclasses.MISSING else {}),
         **data_key_fields(name),
         error_messages=build_error_messages(
-            required_error=required_error, null_error=null_error, invalid_error=invalid_error
+            required_error=required_error, none_error=none_error, invalid_error=invalid_error
         ),
     )
 
@@ -680,7 +680,7 @@ def tuple_field(
     name: str | None = None,
     validate: ValidationFunc | collections.abc.Sequence[ValidationFunc] | None = None,
     required_error: str | None = None,
-    null_error: str | None = None,
+    none_error: str | None = None,
     invalid_error: str | None = None,
     **_: Any,
 ) -> m.fields.Field:
@@ -692,7 +692,7 @@ def tuple_field(
             **default_fields(m.missing),
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -706,7 +706,7 @@ def tuple_field(
             validate=validate,
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -717,7 +717,7 @@ def tuple_field(
         **(default_fields(None) if default is dataclasses.MISSING else {}),
         **data_key_fields(name),
         error_messages=build_error_messages(
-            required_error=required_error, null_error=null_error, invalid_error=invalid_error
+            required_error=required_error, none_error=none_error, invalid_error=invalid_error
         ),
     )
 
@@ -732,7 +732,7 @@ def dict_field(
     name: str | None = None,
     validate: ValidationFunc | collections.abc.Sequence[ValidationFunc] | None = None,
     required_error: str | None = None,
-    null_error: str | None = None,
+    none_error: str | None = None,
     invalid_error: str | None = None,
     **_: Any,
 ) -> m.fields.Field:
@@ -745,7 +745,7 @@ def dict_field(
             **default_fields(m.missing),
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -760,7 +760,7 @@ def dict_field(
             validate=validate,
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -772,7 +772,7 @@ def dict_field(
         **(default_fields(None) if default is dataclasses.MISSING else {}),
         **data_key_fields(name),
         error_messages=build_error_messages(
-            required_error=required_error, null_error=null_error, invalid_error=invalid_error
+            required_error=required_error, none_error=none_error, invalid_error=invalid_error
         ),
     )
 
@@ -786,7 +786,7 @@ def enum_field(
     default: Any = dataclasses.MISSING,
     validate: ValidationFunc | collections.abc.Sequence[ValidationFunc] | None = None,
     required_error: str | None = None,
-    null_error: str | None = None,
+    none_error: str | None = None,
     invalid_error: str | None = None,
     **_: Any,
 ) -> marshmallow.fields.Field:
@@ -798,7 +798,7 @@ def enum_field(
             **default_fields(m.missing),
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -812,7 +812,7 @@ def enum_field(
             validate=validate,
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -823,7 +823,7 @@ def enum_field(
         **(default_fields(None) if default is dataclasses.MISSING else {}),
         **data_key_fields(name),
         error_messages=build_error_messages(
-            required_error=required_error, null_error=null_error, invalid_error=invalid_error
+            required_error=required_error, none_error=none_error, invalid_error=invalid_error
         ),
     )
 
@@ -834,7 +834,7 @@ def raw_field(
     name: str | None = None,
     validate: ValidationFunc | collections.abc.Sequence[ValidationFunc] | None = None,
     required_error: str | None = None,
-    null_error: str | None = None,
+    none_error: str | None = None,
     invalid_error: str | None = None,
     **_: Any,
 ) -> m.fields.Field:
@@ -844,7 +844,7 @@ def raw_field(
         **(default_fields(None) if default is dataclasses.MISSING else {}),
         **data_key_fields(name),
         error_messages=build_error_messages(
-            required_error=required_error, null_error=null_error, invalid_error=invalid_error
+            required_error=required_error, none_error=none_error, invalid_error=invalid_error
         ),
     )
 
@@ -858,7 +858,7 @@ def union_field(
     name: str | None = None,
     validate: ValidationFunc | collections.abc.Sequence[ValidationFunc] | None = None,
     required_error: str | None = None,
-    null_error: str | None = None,
+    none_error: str | None = None,
     invalid_error: str | None = None,
     **_: Any,
 ) -> m.fields.Field:
@@ -870,7 +870,7 @@ def union_field(
             **default_fields(m.missing),
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -884,7 +884,7 @@ def union_field(
             validate=validate,
             **data_key_fields(name),
             error_messages=build_error_messages(
-                required_error=required_error, null_error=null_error, invalid_error=invalid_error
+                required_error=required_error, none_error=none_error, invalid_error=invalid_error
             ),
         )
 
@@ -895,7 +895,7 @@ def union_field(
         **(default_fields(None) if default is dataclasses.MISSING else {}),
         **data_key_fields(name),
         error_messages=build_error_messages(
-            required_error=required_error, null_error=null_error, invalid_error=invalid_error
+            required_error=required_error, none_error=none_error, invalid_error=invalid_error
         ),
     )
 
@@ -914,14 +914,14 @@ UnionField: type[m.fields.Field]
 def build_error_messages(
     *,
     required_error: str | None = None,
-    null_error: str | None = None,
+    none_error: str | None = None,
     invalid_error: str | None = None,
 ) -> dict[str, str] | None:
     error_messages = {}
     if required_error is not None:
         error_messages["required"] = required_error
-    if null_error is not None:
-        error_messages["null"] = null_error
+    if none_error is not None:
+        error_messages["null"] = none_error
     if invalid_error is not None:
         error_messages["invalid"] = invalid_error