+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## v0.0.64a1(2025-11-03)
 
-* Add custom error message support via `required_error`, `null_error`, and `invalid_error` parameters to all field and metadata functions
+* Add custom error message support via `required_error`, `none_error`, and `invalid_error` parameters to all field and metadata functions
 
 
 ## v0.0.63(2025-10-21)